requirements.txt
# FastAPI and dependencies

fastapi==0.115.4

uvicorn[standard]==0.24.0

pydantic==2.12.4

pydantic_core==2.41.5

python-multipart==0.0.20

httpx==0.27.2

# Resilience patterns
tenacity==8.2.3
pybreaker==1.0.1

# Monitoring
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0

# gRPC
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.3

# Database

SQLAlchemy==2.0.44

psycopg2-binary==2.9.11



# Environment and configuration

python-dotenv==1.2.1



# Core dependencies

annotated-types==0.7.0

anyio==4.11.0

click==8.3.1

colorama==0.4.6

greenlet==3.2.4

h11==0.16.0

httptools==0.7.1

idna==3.11

sniffio==1.3.1

starlette==0.40.0

typing_extensions==4.15.0

watchfiles==1.1.1

websockets==15.0.1



# Image processing

Pillow==10.4.0



# Testing

pytest==9.0.1

httpcore==1.0.9



# Additional utilities

packaging==25.0

pluggy==1.6.0

iniconfig==2.3.0
